@@ -14,38 +14,41 @@
   },
   "homepage": "https://www.balistos.com",
   "dependencies": {
-    "axios": "^0.16.1",
-    "prop-types": "^15.5.6",
-    "react": "^15.5.0",
-    "react-dom": "^15.5.0",
-    "react-redux": "^4.4.8",
-    "react-router": "^4.0.0",
-    "react-router-dom": "^4.0.0",
-    "react-slider": "^0.7.0",
-    "react-youtube": "^7.3.0",
-    "redux": "^3.6.0",
-    "redux-saga": "^0.14.4",
-    "socket.io-client": "^1.7.2",
-    "video-time": "^1.1.0"
+    "axios": "0.16.2",
+    "prop-types": "15.5.10",
+    "react": "15.6.1",
+    "react-dom": "15.6.1",
+    "react-redux": "5.0.6",
+    "react-router": "4.1.2",
+    "react-router-dom": "4.1.2",
+    "react-slider": "0.8.0",
+    "react-youtube": "7.4.0",
+    "redux": "3.7.2",
+    "redux-saga": "0.15.6",
+    "socket.io-client": "2.0.3",
+    "video-time": "1.1.0"
   },
   "devDependencies": {
-    "enzyme": "^2.8.0",
-    "eslint": "^3.18.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.3",
-    "node-sass": "^4.5.2",
-    "npm-run-all": "^4.0.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-scripts": "1.0.5",
-    "react-test-renderer": "^15.4.2"
+    "cross-env": "5.0.5",
+    "enzyme": "2.9.1",
+    "eslint": "4.3.0",
+    "eslint-config-airbnb": "15.1.0",
+    "eslint-find-rules": "3.1.1",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-react": "7.1.0",
+    "node-sass": "4.5.3",
+    "npm-run-all": "4.0.2",
+    "react-addons-test-utils": "15.6.0",
+    "react-scripts": "1.0.11",
+    "react-test-renderer": "15.6.1"
   },
   "scripts": {
     "build-css": "node-sass src/components/ -o src/components/",
     "watch-css": "npm run build-css && node-sass src/ -o src/ --watch --recursive",
     "start-js": "react-scripts start",
-    "start": "npm-run-all -p watch-css start-js",
+    "start": "cross-env REACT_APP_API_ENDPOINT=production npm-run-all -p watch-css start-js",
+    "start:local": "cross-env REACT_APP_API_ENDPOINT=local npm-run-all -p watch-css start-js",
     "build": "npm run build-css && react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
