@@ -5,11 +5,11 @@ const CreatePlaylist = (props) => {
   let title;
   let description;
   return (
-    <div className="dropdown" >
+    <div className="dropdown" onClick={(event) => event.stopPropagation()} >
       <form
         onSubmit={(e) => {
           e.preventDefault();
-          if (!title.value.trim() || !description.value.trim) {
+          if (!title.value.trim() || !description.value.trim()) {
             return;
           }
           props.onCreatePlaylistSubmit(title.value, description.value);
@@ -33,4 +33,4 @@ CreatePlaylist.propTypes = {
   onCreatePlaylistSubmit: PropTypes.func.isRequired
 };
 
-module.exports = CreatePlaylist;
+export default CreatePlaylist;
