@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const LogOut = props => (
-  <div className="dropdown small" >
+  <div className="dropdown small" onClick={(event) => event.stopPropagation()} >
     <button className="button green logout" onClick={props.onLogoutClick} >Log Out</button>
   </div> 
 );
@@ -11,4 +11,4 @@ LogOut.propTypes = {
   onLogoutClick: PropTypes.func.isRequired,
 };
 
-module.exports = LogOut;
+export default LogOut;
