@@ -12,9 +12,9 @@ const SearchPlaylistResult = props => (
 SearchPlaylistResult.propTypes = {
   result: PropTypes.shape({
     title: PropTypes.string.isRequired,
-    id: PropTypes.string.isRequired,
+    id: PropTypes.number.isRequired,
     description: PropTypes.string.isRequired,
   }).isRequired,
 };
 
-module.exports = SearchPlaylistResult;
+export default SearchPlaylistResult;
