@@ -1,15 +1,23 @@
 import React from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
+import { connect } from 'react-redux';
 
 import Home from '../Home/Home';
 import Playlist from '../Main/Playlist';
 import Header from '../Header/Header';
+import { closeAllWindows } from '../../actions';
 
 import './App.css';
 
-const App = () => (
-  <BrowserRouter history={history}>
-    <div className="full-height">
+const mapDispatchToProps = (dispatch) => ({
+  clickOutside: () => {
+    dispatch(closeAllWindows());
+  },
+});
+
+const App = (props) => (
+  <BrowserRouter>
+    <div className="full-height" onClick={props.clickOutside}>
       <Header />
       <Route exact path="/" component={Home} />
       <Route path="/playlist/:playlistId" component={Playlist} />
@@ -17,6 +25,5 @@ const App = () => (
   </BrowserRouter>
 );
 
-
-export default App;
+export default connect(undefined, mapDispatchToProps)(App);
 
